# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebDev Orders API",
    "version": "1.0.0",
    "description": "API документация для системы управления заявками на веб-разработку"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.webdev-orders.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Application": {
        "type": "object",
        "required": [
          "title",
          "service_type",
          "contact_email"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Уникальный идентификатор заявки",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID пользователя, создавшего заявку",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "title": {
            "type": "string",
            "description": "Название заявки",
            "example": "Создание корпоративного сайта"
          },
          "description": {
            "type": "string",
            "description": "Описание заявки",
            "example": "Необходимо создать современный корпоративный сайт с CMS"
          },
          "service_type": {
            "type": "string",
            "enum": [
              "landing_page",
              "corporate_site",
              "ecommerce",
              "web_application",
              "other"
            ],
            "description": "Тип услуги",
            "example": "corporate_site"
          },
          "contact_email": {
            "type": "string",
            "format": "email",
            "description": "Контактный email",
            "example": "client@example.com"
          },
          "contact_phone": {
            "type": "string",
            "description": "Контактный телефон",
            "example": 79991234567
          },
          "company_name": {
            "type": "string",
            "description": "Название компании клиента",
            "example": "ООО \"Рога и копыта\""
          },
          "expected_budget": {
            "type": "integer",
            "description": "Ожидаемый бюджет в рублях",
            "example": 150000
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "submitted",
              "in_review",
              "approved",
              "in_progress",
              "completed",
              "cancelled",
              "rejected"
            ],
            "description": "Статус заявки",
            "example": "draft"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "normal",
              "high",
              "urgent"
            ],
            "description": "Приоритет заявки",
            "example": "normal"
          },
          "assigned_to": {
            "type": "string",
            "format": "uuid",
            "description": "ID менеджера, назначенного на заявку",
            "example": "123e4567-e89b-12d3-a456-426614174002"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Дата подачи заявки",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "description": "Крайний срок выполнения",
            "example": "2023-02-01T00:00:00.000Z"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Дата создания заявки",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Дата последнего обновления",
            "example": "2023-01-01T00:00:00.000Z"
          }
        }
      },
      "CreateApplicationRequest": {
        "type": "object",
        "required": [
          "title",
          "service_type",
          "contact_full_name",
          "contact_email",
          "contact_phone"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Название заявки",
            "example": "Создание корпоративного сайта"
          },
          "description": {
            "type": "string",
            "description": "Описание заявки",
            "example": "Необходимо создать современный корпоративный сайт с CMS"
          },
          "service_type": {
            "type": "string",
            "enum": [
              "landing_page",
              "corporate_site",
              "ecommerce",
              "web_application",
              "other"
            ],
            "description": "Тип услуги",
            "example": "corporate_site"
          },
          "contact_full_name": {
            "type": "string",
            "description": "Контактное имя",
            "example": "Иван Иванов"
          },
          "contact_email": {
            "type": "string",
            "format": "email",
            "description": "Контактный email",
            "example": "client@example.com"
          },
          "contact_phone": {
            "type": "string",
            "description": "Контактный телефон",
            "example": 79991234567
          },
          "company_name": {
            "type": "string",
            "description": "Название компании клиента",
            "example": "ООО \"Рога и копыта\""
          },
          "expected_budget": {
            "type": "integer",
            "description": "Ожидаемый бюджет в рублях",
            "example": 150000
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Общее количество элементов",
            "example": 100
          },
          "page": {
            "type": "integer",
            "description": "Номер текущей страницы",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "description": "Количество элементов на странице",
            "example": 10
          },
          "pages": {
            "type": "integer",
            "description": "Общее количество страниц",
            "example": 10
          },
          "has_next": {
            "type": "boolean",
            "description": "Есть ли следующая страница",
            "example": true
          },
          "has_prev": {
            "type": "boolean",
            "description": "Есть ли предыдущая страница",
            "example": false
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "email",
          "full_name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Уникальный идентификатор пользователя",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email пользователя",
            "example": "user@example.com"
          },
          "full_name": {
            "type": "string",
            "description": "Полное имя пользователя",
            "example": "Иван Иванов"
          },
          "phone": {
            "type": "string",
            "description": "Телефон пользователя",
            "example": 79991234567
          },
          "company_name": {
            "type": "string",
            "description": "Название компании",
            "example": "ООО \"Рога и копыта\""
          },
          "role": {
            "type": "string",
            "enum": [
              "client",
              "manager",
              "admin",
              "super_admin"
            ],
            "description": "Роль пользователя в системе",
            "example": "client"
          },
          "is_email_verified": {
            "type": "boolean",
            "description": "Подтвержден ли email",
            "example": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Дата создания пользователя",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Дата последнего обновления",
            "example": "2023-01-01T00:00:00.000Z"
          }
        }
      }
    }
  },
  "paths": {
    "/admin/applications": {
      "get": {
        "summary": "Получение списка заявок с фильтрацией",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Фильтр по статусу заявки"
          },
          {
            "in": "query",
            "name": "service_type",
            "schema": {
              "type": "string"
            },
            "description": "Фильтр по типу услуги"
          },
          {
            "in": "query",
            "name": "priority",
            "schema": {
              "type": "string"
            },
            "description": "Фильтр по приоритету"
          },
          {
            "in": "query",
            "name": "assigned_to",
            "schema": {
              "type": "string"
            },
            "description": "Фильтр по назначенным менеджерам"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Номер страницы"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Количество элементов на странице"
          }
        ],
        "responses": {
          "200": {
            "description": "Список заявок",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "applications": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Application"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/applications": {
      "get": {
        "summary": "Получение списка заявок",
        "tags": [
          "Applications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Фильтр по статусу заявки"
          },
          {
            "in": "query",
            "name": "service_type",
            "schema": {
              "type": "string"
            },
            "description": "Фильтр по типу услуги"
          },
          {
            "in": "query",
            "name": "priority",
            "schema": {
              "type": "string"
            },
            "description": "Фильтр по приоритету"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Номер страницы"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Количество элементов на странице"
          }
        ],
        "responses": {
          "200": {
            "description": "Список заявок",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "applications": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Application"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Создание новой заявки",
        "tags": [
          "Applications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Заявка успешно создана",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "application": {
                          "$ref": "#/components/schemas/Application"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации"
          }
        }
      }
    },
...
    "/auth/register": {
      "post": {
        "summary": "Регистрация нового пользователя",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Email пользователя",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "Пароль пользователя",
                    "example": "password123"
                  },
                  "fullName": {
                    "type": "string",
                    "description": "Полное имя пользователя",
                    "example": "Иван Иванов"
                  },
                  "phone": {
                    "type": "string",
                    "description": "Телефон пользователя",
                    "example": 79991234567
                  },
                  "companyName": {
                    "type": "string",
                    "description": "Название компании",
                    "example": "ООО \"Рога и копыта\""
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь успешно зарегистрирован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        },
                        "tokens": {
                          "type": "object",
                          "properties": {
                            "accessToken": {
                              "type": "string"
                            },
                            "refreshToken": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации или пользователь уже существует"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Вход в систему",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Email пользователя",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "Пароль пользователя",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        },
                        "tokens": {
                          "type": "object",
                          "properties": {
                            "accessToken": {
                              "type": "string"
                            },
                            "refreshToken": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Неверные учетные данные"
          }
        }
      }
    },
...
    "/auth/change-password": {
      "put": {
        "summary": "Изменение пароля",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "description": "Текущий пароль",
                    "example": "oldpassword123"
                  },
                  "newPassword": {
                    "type": "string",
                    "description": "Новый пароль",
                    "example": "newpassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пароль успешно изменен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации"
          }
        }
      }
    },
...
    "/auth/forgot-password": {
      "post": {
        "summary": "Запрос на восстановление пароля",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email пользователя",
                    "example": "user@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Инструкции по восстановлению пароля отправлены",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/applications": {
      "get": {
        "summary": "Получение заявок пользователя",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Фильтр по статусу заявки"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Номер страницы"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Количество элементов на странице"
          }
        ],
        "responses": {
          "200": {
            "description": "Список заявок пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "applications": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Application"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}